Added Dockerfile
Added a Dockerfile to build an image to run puppeteer. Have added a
configuration variable as Chromium throws an error when running as
"root" and not having the "--no-sandbox" flag there.

Updated the README with Docker instructions.

TODO: Instead of installing Chromium through the Debian repos in order
to get the dependencies, we should just install the dependencies by
themselves

TODO: Add config to create a non-root user to run as

TODO: Fix up the file permissions issue where the generated archives
have owner/group set to "root"

Issue(s): None